# lib/api-zod/src/generated/types/createDraftBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDraftBodyMode } from "./createDraftBodyMode";

export interface CreateDraftBody {
  mode: CreateDraftBodyMode;
  title?: string;
  content?: string;
}