# lib/api-zod/src/generated/types/createItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateItemBodyType } from "./createItemBodyType";

export interface CreateItemBody {
  id?: string;
  type: CreateItemBodyType;
  content: string;
  createdAt?: string;
}