# lib/api-zod/src/generated/types/dashboardData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoryWithItems } from "./storyWithItems";

export interface DashboardData {
  totalStories: number;
  activeStories: number;
  archivedStories: number;
  totalItems: number;
  totalDrafts: number;
  recentStories: StoryWithItems[];
}