# lib/api-zod/src/generated/types/draft.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DraftMode } from "./draftMode";

export interface Draft {
  id: string;
  storyId: string;
  mode: DraftMode;
  title: string;
  content: string;
  createdAt: Date;
  updatedAt: Date;
}