# lib/api-zod/src/generated/types/importStoryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImportStoryBodyItemsItem } from "./importStoryBodyItemsItem";

export interface ImportStoryBody {
  title: string;
  items: ImportStoryBodyItemsItem[];
}