# lib/api-zod/src/generated/types/storyWithItems.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoryItem } from "./storyItem";
import type { StoryWithItemsStatus } from "./storyWithItemsStatus";

export interface StoryWithItems {
  id: string;
  title: string;
  status: StoryWithItemsStatus;
  createdAt: Date;
  updatedAt: Date;
  items: StoryItem[];
}